Reshuffle Notifications settings screen
