+++ Sources/App/Resources/en.lproj/Localizable.strings
@@ -73,6 +73,7 @@
 "database.problem.delete" = "Delete Database & Quit App";
 "database.problem.quit" = "Quit App";
 "database.problem.title" = "Database Error";
+"debug_section_label" = "Debug";
 "delete" = "Delete";
 "device.generic_name" = "Device";
 "devices_map.battery_label" = "Battery";
@@ -433,6 +434,7 @@ Your actions and local configuration will still be available after logging in.";
 "settings_details.notifications.categories_synced.footer_no_categories" = "Categories may be also defined in the .yaml configuration.";
 "settings_details.notifications.categories_synced.header" = "Synced Categories";
 "settings_details.notifications.categories.header" = "Categories";
+"settings_details.notifications.categories.footer" = "Actionable notifications are powered by creating categories which define the available actions.";
 "settings_details.notifications.import_legacy_settings.alert.message" = "The push notification categories and actions have been imported from the server.";
 "settings_details.notifications.import_legacy_settings.alert.title" = "Server push configuration imported";
 "settings_details.notifications.import_legacy_settings.button.title" = "Import push configuration from server";
@@ -475,6 +477,7 @@ Your actions and local configuration will still be available after logging in.";
 "settings_details.notifications.sounds.imported_alert.title" = "Sounds Imported";
 "settings_details.notifications.sounds.system" = "System";
 "settings_details.notifications.sounds.title" = "Sounds";
+"settings_details.notifications.sounds.footer" = "Built-in, system or custom sounds can be used with your notifications.";
 "settings_details.notifications.title" = "Notifications";
 "settings_details.notifications.update_section.button.title" = "Update push settings";
 "settings_details.notifications.update_section.footer" = "Updating push settings will request the latest push actions and categories from Home Assistant.";
